requirements.txt
pycardano>=0.11.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
cbor2>=5.6.0
